# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AncesTree",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sample/graph": {
      "get": {
        "tags": [
          "tree"
        ],
        "summary": "Get Sample Graph",
        "description": "A generated, fictional family, for trying the canvas at size. Nothing is stored.",
        "operationId": "get_sample_graph",
        "parameters": [
          {
            "name": "people",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 10,
              "default": 2000,
              "title": "People"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Graph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/layout/positions": {
      "put": {
        "tags": [
          "tree"
        ],
        "summary": "Save Positions",
        "description": "Remember where people were dragged to; a null position puts them back on the rings.",
        "operationId": "save_positions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Positions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tree"
        ],
        "summary": "Clear Positions",
        "description": "Rearrange: everyone goes back to their place on the rings.",
        "operationId": "clear_positions",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "tree"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tree"
        ],
        "summary": "Save Settings",
        "description": "The centre of the tree (none: the oldest ancestor) and how the rings are coloured.",
        "operationId": "save_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exports": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Make Export",
        "description": "Make an export in DATA_DIR/exports: the full archive, GEDCOM 5.5.1, a spreadsheet, or\na copy of the app with the family inside, following its choices: view-only, or to\nedit and send back, for someone.",
        "operationId": "make_export",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{name}": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Download Export",
        "description": "An export or backup from DATA_DIR/exports, as a download.",
        "operationId": "download_export",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backups": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "List Backups",
        "description": "The backup archives, newest first, with what each holds; and the backup folder\n(BACKUP_DIR), with whether it can be reached.",
        "operationId": "list_backups",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Add Backup",
        "description": "Bring in a backup archive from elsewhere, e.g. another disk, to restore from.",
        "operationId": "add_backup",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_backup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backups/{name}/restore": {
      "post": {
        "tags": [
          "export"
        ],
        "summary": "Restore Backup",
        "description": "Replace everything with this backup. Everything as it is now is backed up first.\nRefused (422, \"bad_archive\") if the archive is damaged: then nothing changes.",
        "operationId": "restore_backup",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRestored"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Backup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "made_at": {
            "type": "string",
            "format": "date-time",
            "title": "Made At"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "people": {
            "type": "integer",
            "title": "People"
          },
          "links": {
            "type": "integer",
            "title": "Links"
          },
          "files": {
            "type": "integer",
            "title": "Files"
          },
          "automatic": {
            "type": "boolean",
            "title": "Automatic",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "folder",
          "size",
          "made_at",
          "app_version",
          "people",
          "links",
          "files"
        ],
        "title": "Backup",
        "description": "A backup archive, from its manifest."
      },
      "BackupList": {
        "properties": {
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "reachable": {
            "type": "boolean",
            "title": "Reachable"
          },
          "backups": {
            "items": {
              "$ref": "#/components/schemas/Backup"
            },
            "type": "array",
            "title": "Backups"
          },
          "automatic_backups": {
            "type": "boolean",
            "title": "Automatic Backups",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "folder",
          "reachable",
          "backups"
        ],
        "title": "BackupList"
      },
      "BackupRestored": {
        "properties": {
          "made_at": {
            "type": "string",
            "format": "date-time",
            "title": "Made At"
          },
          "people": {
            "type": "integer",
            "title": "People"
          },
          "links": {
            "type": "integer",
            "title": "Links"
          },
          "files": {
            "type": "integer",
            "title": "Files"
          },
          "backup": {
            "type": "string",
            "title": "Backup"
          }
        },
        "type": "object",
        "required": [
          "made_at",
          "people",
          "links",
          "files",
          "backup"
        ],
        "title": "BackupRestored"
      },
...
      "Body_add_backup": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "An AncesTree backup archive (.zip)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_add_backup"
      },
...
      "CopyPermissions": {
        "properties": {
          "add": {
            "type": "boolean",
            "title": "Add",
            "default": true
          },
          "change": {
            "type": "boolean",
            "title": "Change",
            "default": true
          },
          "remove": {
            "type": "boolean",
            "title": "Remove",
            "default": true
          },
          "stories": {
            "type": "boolean",
            "title": "Stories",
            "default": true
          },
          "photos": {
            "type": "boolean",
            "title": "Photos",
            "default": true
          }
        },
        "type": "object",
        "title": "CopyPermissions",
        "description": "What relatives may do in a copy to edit: all of it unless switched off."
      },
...
      "DateQualifier": {
        "type": "string",
        "enum": [
          "exact",
          "about",
          "before",
          "after",
          "between"
        ],
        "title": "DateQualifier"
      },
...
      "ExportFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "name",
          "folder",
          "format",
          "size"
        ],
        "title": "ExportFile"
      },
      "ExportFormat": {
        "type": "string",
        "enum": [
          "archive",
          "gedcom",
          "csv",
          "copy"
        ],
        "title": "ExportFormat"
      },
      "ExportRequest": {
        "properties": {
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "title": {
            "type": "string",
            "maxLength": 60,
            "title": "Title",
            "default": ""
          },
          "hide_living": {
            "type": "boolean",
            "title": "Hide Living",
            "default": false
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "archive": {
            "type": "boolean",
            "title": "Archive",
            "default": true
          },
          "editable": {
            "type": "boolean",
            "title": "Editable",
            "default": false
          },
          "for_name": {
            "type": "string",
            "maxLength": 60,
            "title": "For Name",
            "default": ""
          },
          "may": {
            "$ref": "#/components/schemas/CopyPermissions"
          }
        },
        "type": "object",
        "required": [
          "format"
        ],
        "title": "ExportRequest"
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "unknown"
        ],
        "title": "Gender"
      },
...
      "Graph": {
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/GraphPerson"
            },
            "type": "array",
            "title": "People"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/GraphLink"
            },
            "type": "array",
            "title": "Links"
          },
          "layout": {
            "$ref": "#/components/schemas/GraphLayout"
          }
        },
        "type": "object",
        "required": [
          "people",
          "links",
          "layout"
        ],
        "title": "Graph"
      },
      "GraphLayout": {
        "properties": {
          "units": {
            "items": {
              "$ref": "#/components/schemas/LayoutUnit"
            },
            "type": "array",
            "title": "Units"
          },
          "seats": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Seat"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "Seats"
          },
          "unlinked": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Unlinked"
          },
          "centre_chosen": {
            "type": "boolean",
            "title": "Centre Chosen"
          }
        },
        "type": "object",
        "required": [
          "units",
          "seats",
          "unlinked",
          "centre_chosen"
        ],
        "title": "GraphLayout"
      },
      "GraphLink": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "parent",
              "spouse"
            ],
            "title": "Type"
          },
          "source": {
            "type": "string",
            "format": "uuid",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "format": "uuid",
            "title": "Target"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpouseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "in_layout": {
            "type": "boolean",
            "title": "In Layout",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "source",
          "target",
          "kind",
          "status"
        ],
        "title": "GraphLink",
        "description": "For \"parent\" links, `source` is the parent and `target` the child."
      },
      "GraphPerson": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "death_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Year"
          },
          "birth_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Order"
          },
          "placeholder": {
            "type": "boolean",
            "title": "Placeholder"
          },
          "photo_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Version"
          },
          "birthplace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthplace"
          },
          "born_in": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Born In"
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "born": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "died": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_living": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Living"
          },
          "birth_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "nickname",
          "gender",
          "birth_year",
          "death_year",
          "birth_order",
          "placeholder",
          "photo_version"
        ],
        "title": "GraphPerson"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LayoutUnit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "centre": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Centre"
          },
          "anchor": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor"
          },
          "anchor_seat": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seat"
              },
              {
                "type": "null"
              }
            ]
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "id",
          "centre",
          "anchor",
          "anchor_seat",
          "size"
        ],
        "title": "LayoutUnit",
        "description": "One set of rings: a family's own, or a cluster hanging off one person."
      },
...
      "PartialDate": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "qualifier": {
            "$ref": "#/components/schemas/DateQualifier",
            "default": "exact"
          },
          "year_to": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year To"
          },
          "original_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          }
        },
        "type": "object",
        "title": "PartialDate",
        "description": "A date that may be incomplete or approximate: \"1950\", \"Mar 1950\", \"c. 1920\", \"1910-1915\".\n\nUnknown parts stay empty; they are never filled with a stand-in value."
      },
...
      "Position": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "id",
          "x",
          "y"
        ],
        "title": "Position"
      },
      "Positions": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/Position"
            },
            "type": "array",
            "maxItems": 20000,
            "title": "Positions"
          }
        },
        "type": "object",
        "required": [
          "positions"
        ],
        "title": "Positions"
      },
...
      "SpouseStatus": {
        "type": "string",
        "enum": [
          "married",
          "divorced",
          "widowed"
        ],
        "title": "SpouseStatus"
      },
...
      "TreeSettings": {
        "properties": {
          "centre": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Centre"
          },
          "colours": {
            "type": "string",
            "enum": [
              "branch",
              "generation",
              "closeness",
              "off"
            ],
            "title": "Colours",
            "default": "branch"
          }
        },
        "type": "object",
        "title": "TreeSettings",
        "description": "Stored in DATA_DIR/settings/app.json."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}